Refactoring,  split single js and python file in several files
