+++ rg_instructor_analytics/static/rg_instructor_analytics/js/Base.js
@@ -1,0 +1,50 @@
+
+
+$(function () {
+    // import Tab from 'Tab';
+    // import CohortTab from 'CohortTab';
+    // import EnrollmentTab from 'EnrollmentTab';
+    // import GradebookTab from 'GradebookTab';
+    // import ProblemTab from 'ProblemTab';
+
+    'use strict';
+    var CSS_INSTRUCTOR_CONTENT = 'instructor-dashboard-content-2';
+
+    var tabList = [];
+
+    function toggleToTab(tab) {
+        tabList.forEach(function (t) {
+            t.setActive(t === tab);
+        });
+    }
+
+    $(function () {
+        var $content = $('.' + CSS_INSTRUCTOR_CONTENT);
+        tabList = [
+            EnrollmentTab(
+                $content.find('#enrollment-stats-btn'),
+                $content.find('#section-enrollment-stats')),
+            ProblemTab(
+                $content.find('#problems-btn'),
+                $content.find('#section-problem')),
+            GradebookTab(
+                $content.find('#gradebook-btn'),
+                $content.find('#section-gradebook')),
+            CohortTab(
+                $content.find('#cohort-btn'),
+                $content.find('#section-cohort'))
+        ];
+        tabList.forEach(function (tab) {
+            tab.button.click(function () {
+                toggleToTab(tab)
+            })
+        });
+
+        toggleToTab(tabList[0]);
+    });
+
+    window.setup_debug = function (element_id, edit_link, staff_context) {
+        // stub function.
+    };
+
+}).call(this);
