+++ rg_instructor_analytics/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+Util package.
+"""
