+++ rg_instructor_analytics/views/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package with views.
+"""
